requirements.txt
torch
torchvision
torchaudio
opencv-python
matplotlib
tqdm
imageio
transformers
accelerate
pillow
git+https://github.com/facebookresearch/segment-anything.git
